server/requirements.txt
openenv-core[core]
fastapi
uvicorn
openai
requests
numpy
pydantic
